refactor: mirror core feature test folders
